+++ tests/Integration/Defer/DeferDBTest.php
@@ -2,7 +2,6 @@
 
 namespace Tests\Integration\Defer;
 
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection as BaseCollection;
 use Illuminate\Support\Facades\DB;
@@ -37,9 +36,7 @@ class DeferDBTest extends DBTestCase
             'company_id' => $company->getKey(),
         ]);
 
-        $this->mockResolver(function () use ($user): User {
-            return $user;
-        });
+        $this->mockResolver($user);
 
         $this->schema = /** @lang GraphQL */ '
         type Company {
@@ -92,9 +89,7 @@ class DeferDBTest extends DBTestCase
         ]);
         $user = $users[0];
 
-        $this->mockResolver(function () use ($user): User {
-            return $user;
-        });
+        $this->mockResolver($user);
 
         $this->schema = /** @lang GraphQL */ '
         type Company {
@@ -168,9 +163,7 @@ class DeferDBTest extends DBTestCase
                 ]);
             });
 
-        $this->mockResolver(function () use ($companies): Collection {
-            return $companies;
-        });
+        $this->mockResolver($companies);
 
         $this->schema = /** @lang GraphQL */ '
         type Company {