+++ tests/Integration/Defer/DeferTest.php
@@ -13,11 +13,6 @@ class DeferTest extends TestCase
 {
     use SetUpDefer;
 
-    /**
-     * @var mixed[]
-     */
-    public static $data = [];
-
     protected function getEnvironmentSetUp($app)
     {
         parent::getEnvironmentSetUp($app);
@@ -55,25 +50,25 @@ class DeferTest extends TestCase
 
     public function testCanDeferFields(): void
     {
-        self::$data = [
+        $this->mockResolver([
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
             ],
-        ];
+        ]);
 
-        $this->schema = "
+        $this->schema = /** @lang GraphQL */ '
         type User {
             name: String!
             parent: User
         }
 
         type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
+            user: User @mock
         }
-        ";
+        ';
 
-        $chunks = $this->getStreamedChunks('
+        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
         {
             user {
                 name
@@ -108,7 +103,7 @@ class DeferTest extends TestCase
 
     public function testCanDeferNestedFields(): void
     {
-        self::$data = [
+        $data = [
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
@@ -117,19 +112,20 @@ class DeferTest extends TestCase
                 ],
             ],
         ];
+        $this->mockResolver($data);
 
-        $this->schema = "
+        $this->schema = /** @lang GraphQL */ '
         type User {
             name: String!
             parent: User
         }
 
         type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
+            user: User @mock
         }
-        ";
+        ';
 
-        $chunks = $this->getStreamedChunks('
+        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
         {
             user {
                 name
@@ -145,46 +141,45 @@ class DeferTest extends TestCase
 
         $this->assertCount(3, $chunks);
 
-        $this->assertSame(self::$data['name'], Arr::get($chunks[0], 'data.user.name'));
+        $this->assertSame($data['name'], Arr::get($chunks[0], 'data.user.name'));
         $this->assertNull(Arr::get($chunks[0], 'data.user.parent'));
 
         $deferred = Arr::get($chunks[1], 'user.parent');
         $this->assertArrayHasKey('name', $deferred['data']);
-        $this->assertSame(self::$data['parent']['name'], $deferred['data']['name']);
+        $this->assertSame($data['parent']['name'], $deferred['data']['name']);
         $this->assertArrayHasKey('parent', $deferred['data']);
         $this->assertNull($deferred['data']['parent']);
 
         $nestedDeferred = Arr::get($chunks[2], 'user.parent.parent');
         $this->assertArrayHasKey('name', $nestedDeferred['data']);
-        $this->assertSame(self::$data['parent']['parent']['name'], $nestedDeferred['data']['name']);
+        $this->assertSame($data['parent']['parent']['name'], $nestedDeferred['data']['name']);
     }
 
     public function testCanDeferNestedFieldsOnMutations(): void
     {
-        self::$data = [
+        $this->mockResolver([
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
             ],
-        ];
+        ]);
 
-        $this->schema = /** @lang GraphQL */ "
+        $this->schema = /** @lang GraphQL */ '
         type User {
             name: String!
             parent: User
         }
 
         type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
+            user: User @mock
         }
 
         type Mutation {
-            updateUser(name: String!): User
-                @field(resolver: \"{$this->qualifyTestResolver()}\")
+            updateUser(name: String!): User @mock
         }
-        ";
+        ';
 
-        $chunks = $this->getStreamedChunks('
+        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
         mutation {
             updateUser(
                 name: "Foo"
@@ -221,7 +216,7 @@ class DeferTest extends TestCase
 
     public function testCanDeferListFields(): void
     {
-        self::$data = [
+        $data = [
             [
                 'title' => 'Foo',
                 'author' => [
@@ -235,8 +230,9 @@ class DeferTest extends TestCase
                 ],
             ],
         ];
+        $this->mockResolver($data);
 
-        $this->schema = /** @lang GraphQL */ "
+        $this->schema = /** @lang GraphQL */ '
         type Post {
             title: String
             author: User
@@ -247,11 +243,11 @@ class DeferTest extends TestCase
         }
 
         type Query {
-            posts: [Post] @field(resolver: \"{$this->qualifyTestResolver()}\")
+            posts: [Post] @mock
         }
-        ";
+        ';
 
-        $chunks = $this->getStreamedChunks('
+        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
         {
             posts {
                 title
@@ -268,15 +264,15 @@ class DeferTest extends TestCase
         $this->assertNull(Arr::get($chunks[0], 'data.posts.1.author'));
 
         $deferredPost1 = $chunks[1]['posts.0.author']['data'];
-        $this->assertSame(self::$data[0]['author']['name'], Arr::get($deferredPost1, 'name'));
+        $this->assertSame($data[0]['author']['name'], Arr::get($deferredPost1, 'name'));
 
         $deferredPost2 = $chunks[1]['posts.1.author']['data'];
-        $this->assertSame(self::$data[1]['author']['name'], Arr::get($deferredPost2, 'name'));
+        $this->assertSame($data[1]['author']['name'], Arr::get($deferredPost2, 'name'));
     }
 
     public function testCanDeferGroupedListFields(): void
     {
-        self::$data = [
+        $data = [
             [
                 'title' => 'Foo',
                 'author' => [
@@ -296,8 +292,9 @@ class DeferTest extends TestCase
                 ],
             ],
         ];
+        $this->mockResolver($data);
 
-        $this->schema = /** @lang GraphQL */ "
+        $this->schema = /** @lang GraphQL */ '
         type Comment {
             message: String
         }
@@ -313,11 +310,11 @@ class DeferTest extends TestCase
         }
 
         type Query {
-            posts: [Post] @field(resolver: \"{$this->qualifyTestResolver()}\")
+            posts: [Post] @mock
         }
-        ";
+        ';
 
-        $chunks = $this->getStreamedChunks('
+        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
         {
             posts {
                 title
@@ -337,39 +334,23 @@ class DeferTest extends TestCase
         $this->assertNull(Arr::get($chunks[0], 'data.posts.1.author'));
 
         $deferredPost1 = $chunks[1]['posts.0.author']['data'];
-        $this->assertSame(self::$data[0]['author']['name'], Arr::get($deferredPost1, 'name'));
+        $this->assertSame($data[0]['author']['name'], Arr::get($deferredPost1, 'name'));
 
         $deferredComment1 = $chunks[1]['posts.0.comments']['data'];
         $this->assertCount(1, $deferredComment1);
-        $this->assertSame(self::$data[0]['comments'][0]['message'], Arr::get($deferredComment1[0], 'message'));
+        $this->assertSame($data[0]['comments'][0]['message'], Arr::get($deferredComment1[0], 'message'));
 
         $deferredPost2 = $chunks[1]['posts.1.author']['data'];
-        $this->assertSame(self::$data[1]['author']['name'], Arr::get($deferredPost2, 'name'));
+        $this->assertSame($data[1]['author']['name'], Arr::get($deferredPost2, 'name'));
 
         $deferredComment2 = $chunks[1]['posts.1.comments']['data'];
         $this->assertCount(1, $deferredComment2);
-        $this->assertSame(self::$data[1]['comments'][0]['message'], Arr::get($deferredComment2[0], 'message'));
+        $this->assertSame($data[1]['comments'][0]['message'], Arr::get($deferredComment2[0], 'message'));
     }
 
     public function testCancelsDefermentAfterMaxExecutionTime(): void
     {
-        $this->schema = "
-        type User {
-            name: String!
-            parent: User
-        }
-
-        type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
-        }
-        ";
-
-        /** @var \Nuwave\Lighthouse\Defer\Defer $defer */
-        $defer = app(Defer::class);
-        // Set max execution time to now so we immediately resolve deferred fields
-        $defer->setMaxExecutionTime(microtime(true));
-
-        self::$data = [
+        $data = [
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
@@ -378,8 +359,25 @@ class DeferTest extends TestCase
                 ],
             ],
         ];
+        $this->mockResolver($data);
 
-        $chunks = $this->getStreamedChunks('
+        $this->schema = /** @lang GraphQL */ '
+        type User {
+            name: String!
+            parent: User
+        }
+
+        type Query {
+            user: User @mock
+        }
+        ';
+
+        /** @var \Nuwave\Lighthouse\Defer\Defer $defer */
+        $defer = app(Defer::class);
+        // Set max execution time to now so we immediately resolve deferred fields
+        $defer->setMaxExecutionTime(microtime(true));
+
+        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
         {
             user {
                 name
@@ -396,34 +394,19 @@ class DeferTest extends TestCase
         // If we didn't hit the max execution time we would have 3 items in the array
         $this->assertCount(2, $chunks);
 
-        $this->assertSame(self::$data['name'], Arr::get($chunks[0], 'data.user.name'));
+        $this->assertSame($data['name'], Arr::get($chunks[0], 'data.user.name'));
         $this->assertNull(Arr::get($chunks[0], 'data.user.parent'));
 
         $deferred = Arr::get($chunks[1], 'user.parent');
         $this->assertArrayHasKey('name', $deferred['data']);
-        $this->assertSame(self::$data['parent']['name'], $deferred['data']['name']);
+        $this->assertSame($data['parent']['name'], $deferred['data']['name']);
         $this->assertArrayHasKey('parent', $deferred['data']);
-        $this->assertSame(self::$data['parent']['parent']['name'], $deferred['data']['parent']['name']);
+        $this->assertSame($data['parent']['parent']['name'], $deferred['data']['parent']['name']);
     }
 
     public function testCancelsDefermentAfterMaxNestedFields(): void
     {
-        $this->schema = "
-        type User {
-            name: String!
-            parent: User
-        }
-
-        type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
-        }
-        ";
-
-        /** @var \Nuwave\Lighthouse\Defer\Defer $defer */
-        $defer = app(Defer::class);
-        $defer->setMaxNestedFields(1);
-
-        self::$data = [
+        $data = [
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
@@ -432,8 +415,24 @@ class DeferTest extends TestCase
                 ],
             ],
         ];
+        $this->mockResolver($data);
 
-        $chunks = $this->getStreamedChunks('
+        $this->schema = /** @lang GraphQL */ '
+        type User {
+            name: String!
+            parent: User
+        }
+
+        type Query {
+            user: User @mock
+        }
+        ';
+
+        /** @var \Nuwave\Lighthouse\Defer\Defer $defer */
+        $defer = app(Defer::class);
+        $defer->setMaxNestedFields(1);
+
+        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
         {
             user {
                 name
@@ -449,35 +448,35 @@ class DeferTest extends TestCase
 
         $this->assertCount(2, $chunks);
 
-        $this->assertSame(self::$data['name'], Arr::get($chunks[0], 'data.user.name'));
+        $this->assertSame($data['name'], Arr::get($chunks[0], 'data.user.name'));
         $this->assertNull(Arr::get($chunks[0], 'data.user.parent'));
 
         $deferred = Arr::get($chunks[1], 'user.parent');
         $this->assertArrayHasKey('name', $deferred['data']);
-        $this->assertSame(self::$data['parent']['name'], $deferred['data']['name']);
+        $this->assertSame($data['parent']['name'], $deferred['data']['name']);
         $this->assertArrayHasKey('parent', $deferred['data']);
-        $this->assertSame(self::$data['parent']['parent']['name'], $deferred['data']['parent']['name']);
+        $this->assertSame($data['parent']['parent']['name'], $deferred['data']['parent']['name']);
     }
 
     public function testThrowsExceptionOnNunNullableFields(): void
     {
-        self::$data = [
+        $this->mockResolver([
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
             ],
-        ];
+        ]);
 
-        $this->schema = "
+        $this->schema = /** @lang GraphQL */ '
         type User {
             name: String!
             parent: User!
         }
 
         type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
+            user: User @mock
         }
-        ";
+        ';
 
         $this->graphQL(/** @lang GraphQL */ '
         {
@@ -499,7 +498,7 @@ class DeferTest extends TestCase
 
     public function testDoesNotDeferWithIncludeAndSkipDirectives(): void
     {
-        self::$data = [
+        $this->mockResolver([
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
@@ -507,9 +506,9 @@ class DeferTest extends TestCase
                     'name' => 'Mr. Smith',
                 ],
             ],
-        ];
+        ]);
 
-        $this->schema = "
+        $this->schema = /** @lang GraphQL */ '
         directive @include(if: Boolean!) on FIELD
         directive @skip(if: Boolean!) on FIELD
 
@@ -519,9 +518,9 @@ class DeferTest extends TestCase
         }
 
         type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
+            user: User @mock
         }
-        ";
+        ';
 
         $this->graphQL(/** @lang GraphQL */ '
         {
@@ -552,23 +551,24 @@ class DeferTest extends TestCase
 
     public function testRequiresDeferDirectiveOnAllFieldDeclarations(): void
     {
-        self::$data = [
+        $data = [
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
             ],
         ];
+        $this->mockResolver($data);
 
-        $this->schema = "
+        $this->schema = /** @lang GraphQL */ '
         type User {
             name: String!
             parent: User
         }
 
         type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
+            user: User @mock
         }
-        ";
+        ';
 
         $this->graphQL(/** @lang GraphQL */ '
         fragment UserWithParent on User {
@@ -587,32 +587,27 @@ class DeferTest extends TestCase
         }
         ')->assertJson([
             'data' => [
-                'user' => self::$data,
+                'user' => $data,
             ],
         ]);
     }
 
     public function testThrowsIfTryingToDeferRootMutationFields(): void
     {
-        self::$data = [
-            'name' => 'John Doe',
-        ];
+        $this->mockResolverExpects(
+            $this->never()
+        );
 
-        $this->schema = /** @lang GraphQL */ "
+        $this->schema = /** @lang GraphQL */ '
         type User {
             name: String!
             parent: User
         }
 
-        type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
-        }
-
         type Mutation {
-            updateUser(name: String!): User
-                @field(resolver: \"{$this->qualifyTestResolver()}\")
+            updateUser(name: String!): User @mock
         }
-        ";
+        '.self::PLACEHOLDER_QUERY;
 
         $this->graphQL(/** @lang GraphQL */ '
         mutation UpdateUser {
@@ -631,23 +626,24 @@ class DeferTest extends TestCase
 
     public function testDoesNotDeferFieldsIfFalse(): void
     {
-        self::$data = [
+        $data = [
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
             ],
         ];
+        $this->mockResolver($data);
 
-        $this->schema = /** @lang GraphQL */ "
+        $this->schema = /** @lang GraphQL */ '
         type User {
             name: String!
             parent: User
         }
 
         type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
+            user: User @mock
         }
-        ";
+        ';
 
         $this->graphQL(/** @lang GraphQL */ '
         {
@@ -660,32 +656,39 @@ class DeferTest extends TestCase
         }
         ')->assertJson([
             'data' => [
-                'user' => self::$data,
+                'user' => $data,
             ],
         ]);
     }
 
     public function testIncludesErrorsForDeferredFields(): void
     {
-        self::$data = [
+        $this->mockResolver([
             'name' => 'John Doe',
             'parent' => [
                 'name' => 'Jane Doe',
             ],
-        ];
+        ]);
 
-        $this->schema = /** @lang GraphQL */ "
+        $this->mockResolver(
+            function () {
+                throw new Error('deferred_exception');
+            },
+            'throw'
+        );
+
+        $this->schema = /** @lang GraphQL */ '
         type User {
             name: String!
-            parent: User @field(resolver: \"{$this->qualifyTestResolver('throw')}\")
+            parent: User @mock(key: "throw")
         }
 
         type Query {
-            user: User @field(resolver: \"{$this->qualifyTestResolver()}\")
+            user: User @mock
         }
-        ";
+        ';
 
-        $chunks = $this->getStreamedChunks('
+        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
         {
             user {
                 name
@@ -703,18 +706,5 @@ class DeferTest extends TestCase
         $this->assertNull($parent['user.parent']['data']);
         $this->assertArrayHasKey('errors', $parent['user.parent']);
         $this->assertCount(1, $parent['user.parent']['errors']);
-    }
-
-    public function resolve(): array
-    {
-        return self::$data;
-    }
-
-    /**
-     * @throws \GraphQL\Error\Error
-     */
-    public function throw(): void
-    {
-        throw new Error('deferred_exception');
     }
 }