+++ CHANGELOG.md
@@ -19,6 +19,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
   with Lumen https://github.com/nuwave/lighthouse/pull/1100
 - Add test trait `\Nuwave\Lighthouse\Testing\UsesTestSchema` to enable using
   a dummy schema for testing custom Lighthouse extensions https://github.com/nuwave/lighthouse/pull/1171
+- Simplify mocking resolvers that just return static data https://github.com/nuwave/lighthouse/pull/1177
 
 ### Fixed